cells: make InversionChecker; handle all size data
Now inversion encoding also prepends a length field, and it should now
support input cell data of any length via padding. It passes the unit
tests in inversion.py - still trying to get it to work in dcnet.py.

This commit also splits the verify function into a superclass of
InversionEncoder, InversionChecker. It is now called checking instead of
verifying because we already have "verify" and "certify" for different
purposes.

Finally, this commit adds the very short "debug" function to utils, for
easy toggling of debug statements.