cells: Add request cell encoder
The idea is to add a single, ownerless, k-bit "request cell" to the
transmission schedule. The request cell is a Bloom filter representing
the set of clients requesting a slot. A slot owner wishing to write data
in the next round transmits a unique bit sequence in the request cell.
At the end of the interval, trustees reveal their trap secrets as usual
and use this to compute the OR of all client hashes. Any bits that
remain 0 (i.e., bits that no client legitimately requesting a slot would
have set) constitute trap bits.
This commit's code implements the bloom filter and trap checking part of
this design. It passes its unit tests but does not yet deal with the
possibility of false negatives introduced by the DC-nets.