cells: Add parameter optimization to request
Adds an Optimizer class for efficiently computing the necessary bit
length and the method for deciding on the bloom filter hashes given
desired properties (currently probability of hash collisions, and
probability of an arbitrary bit being a trap bit). I am still working on
a write up of the math behind what is in this commit.
Also adds a function for reporting the request cell size to the outside
world. With the parameters currently in the file, it matches the cell
size of other cells at 24.