dcnet,certify: Store nym-indexed text as dicts
Now everything that is indexed by slot owner index is a dictionary
instead of a list. This reflects the fact that some slots, in particular
the request slot, will not correspond to a single nym. Names of
variables that really corespond to slots instead of nyms have been
changed to reflect that.