dcnet,cells: Integrate request cell with dcnet
Now each round's ciphertext/cleartext has a slot for a request cell in
addition to the slot for each nym. All clients and trustees write xornet
ciphertext to this cell, and slot-owners requesting additional space
also write their request code. The relay accumulates the requests it
receives and sends the accumulated request back to the clients in its
cleartext each round. At the end of the interval, trustees reveal the
trap secrets and verify the validity of the request cells. At this
point, the trap secrets can be sent to the relay so that the relay can
read the bloom filter and come up with the list of nyms requesting
slots.