cells,dcnet: Make request retries probabilistic
Now, when a client discovers its requst missing from the last round's
cleartext, it only attempts to retransmit the missing 1 bits. For each
missing 1 bit, it randomly sends a 1 or a 0 so that over successive
rounds, with high probability, exactly one client will write a 1 to that
bit and the relay will get the request.

The probabilistic logic is in request.py, which now takes an existing
cell as an argument to encode. The Client class in dcnet.py keeps track
of how many times each nym has unsuccessfully attempted a request.

The test now consistently terminates successfully within 10 rounds.