Deployment works on Zoo
Deployment currently works on the Zoo. You have to get a little bit
lucky in terms of port usage. Sometimes other people will be using
the ports that you need. You can see this when you see a "tcp bind"
error show up in the log messages. This code has to be cleaned up a
little bit to become more reliable.

This deployment process should work the same anywhere that remote ssh
key pairs have been set up with, as long as hosts are either provided
in the command line or in the configuration file itself.

Next on the horizon is that we need to be able to specify a specific
username to use when logging into machines. In addition, the configuration
file and the deployment tool should add a option for just renaming the
ports of the configuration file. This way hostnames can remane fixed
but if port usage is not working out (or old ports have not been freed)
then we can change the ports that we are using rather than waiting for
the OS to garbage collect them.

We also need to add a way for the deployment script to learn when the
collective signing nodes have completed and whether they were successful.