added in support for usernames and port rewriting.
Now you can specify your username for logging into hosts. You can
also rewrite the ports in the config file. There is a simple timer
that runs during announcing. Currently for a system of 6 nodes it
takes 0.04 seconds per announcment round. We will have to test this
on planet lab for a more realistic test.