deploy2pl prototype finished
Takes in a list of hostnames (could be planetlab) and then runs the
latency_test script on each of the machines to determine its speed.
It then creates a tree of a user specified depth (2 by default), and
runs the exec script (signing nodes) on each of these servers with
this generated configuration.

The script needs to be altered slightly to add timeouts for the initial
scp phases. If a node is too slow it should be considered failed. In addition
a threshold should be put in place for pinging times. Currently it only
does not use a node if it fails to ping it, but instead it should not use
a node if it is "too" slow.