deploying to nodes
fixed error with mutex rlock in tcphost not being released on
error within one of its children. it was causing lack of progress.
currently there is a bug being exhibited where the root node is not
connecting to its children properly. thus sometimes it exhibits (very
quick) rounds in which it doesn't communicate with its children. the bug
might have to do with the exponential backoff when connecting (in
oldconfig). Sometimes the bug also manifests itself as a series of
timeouts leading to EOF and then program termination (desired on EOF).
This bug only appears when running on large host configs 12 nodes per
machine. also care should be taken that it is isn't due to an assumption
that we have made by hardcoding wait times in. also it should be noted
that consistently, when we exceed the number of processors on the
machine, many problems are experienced. we should make sure that this
isn't due to some deterlab nuance.