Testing on thousands of hosts
Distributed Logserver across three different machines due to file
limits. Replica loggers forward their log entries to the master logger.
This allows for the tcp connections clients make to be distributed
across multiple loggers. Currently the number of loggers is set to 3,
but this should be easily adjusted if we add in more machines to be
loggers.