fixed bug in viewchange protocol
viewchange protocol had a race condition, where multiple viewchanges
could be initiated within the same round. this resulted in the nodes
having inconsistent state between them, where some are waiting for a
viewchange and others aren't. this was solved by "skipping" the
viewchange round. if a viewchange is scheduled for round r, initiate the
view change, and then advance to round r+1 (propogating this new round
down from the root). this means that, if all nodes are functioning
properly, there will be consensus on the new round and the new leader.